weather_monitoring/requirements.txt
apache-airflow == 2.8.0
kafka-python
spark == 3.5.0
pyspark
python-dotenv
cassandra-driver